requirements.txt
requests>=2.28
terminaltables>=3.1
python-dotenv>=1.0